Move formatter code and tests into internal/format package
Relocate the formatter implementation (modelicafmt.go, template.go) and its
tests into a dedicated internal/format package, leaving a thin package main
(main.go) at the repository root as the CLI entry point. The bundled test data
moves with the package to internal/format/testdata/ (the idiomatic Go location,
next to the tests that use it).

To let package main drive the formatter across the package boundary:
- export ProcessFile (was processFile) and give it an explicit dialect
  parameter instead of reading the -template flag global directly, removing the
  core->CLI coupling
- add a NewConfig constructor so callers can build a Config without the
  unexported fields being exported
- export DialectJinja (was dialectJinja) for the -template flag default

The tests stay white-box in package format, so they keep full access to the
unexported internals with no API churn. Update update_testdata.sh and the README
usage snippet to the new testdata path.
