+++ internal/format/testdata/example-arrays-nd-out.mo
