+++ internal/format/testdata/example-arrays-nd.mo
