+++ internal/format/testdata/example-arrays-out.mo
