+++ internal/format/testdata/example-arrays-wrapped-out.mo
