+++ internal/format/testdata/example-arrays.mo
