+++ internal/format/testdata/example-no-within-out.mo
