+++ internal/format/testdata/example-no-within.mo
