+++ internal/format/testdata/functions-out.mo
