+++ internal/format/testdata/gmt-boiler-polynomial-out.mot
