+++ internal/format/testdata/gmt-boiler-polynomial.mot
