+++ internal/format/testdata/gmt-building-80-out.mo
