+++ internal/format/testdata/gmt-building-empty-lines-out.mo
