+++ internal/format/testdata/gmt-building-out.mo
