+++ internal/format/testdata/gmt-building.mo
