+++ internal/format/testdata/gmt-cooling-indirect-out.mot
