+++ internal/format/testdata/gmt-cooling-indirect.mot
