+++ internal/format/testdata/gmt-coolingtower-out.mo
