+++ internal/format/testdata/gmt-coolingtower.mo
