+++ internal/format/testdata/gmt-design-data-series-out.mot
