+++ internal/format/testdata/gmt-design-data-series.mot
