+++ internal/format/testdata/gmt-district-energy-system.mot
