+++ internal/format/testdata/gmt-hptrio-variable-dist.mot
