+++ internal/format/testdata/gmt-run-spawn-building.mot
