+++ internal/format/testdata/gmt-spawn-building-out.mot
