+++ internal/format/testdata/gmt-spawn-building.mot
