+++ internal/format/testdata/gmt-time-series-building-out.mot
