+++ internal/format/testdata/gmt-time-series-building.mot
